pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mda-osint-db"
version = "0.1.0"
description = "Maritime Domain Awareness — Open-Source Intelligence Database"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "MDA Community" },
]
keywords = ["maritime", "domain-awareness", "osint", "intelligence", "graph-database"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Security",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "asyncpg>=0.29.0",
    "gqlalchemy>=1.6.0",
    "kafka-python>=2.0.2",
    "pydantic>=2.6.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "requests>=2.31.0",
    "rdflib>=7.0.0",
    "splink>=3.9.0",
    "elasticsearch[async]>=8.12.0",
    "minio>=7.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
nlp = [
    "spacy>=3.7.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
